% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_analysis.R
\name{annotate_seq_tbl}
\alias{annotate_seq_tbl}
\title{Amplicon haplotype inference}
\usage{
annotate_seq_tbl(
  seq_tbl,
  marker_info,
  output_dir = NULL,
  threads = 8L,
  min_marker_count = 100L,
  min_asv_count = 5L,
  min_asv_freq = 0.001,
  min_ident = 0.75,
  min_ident_z = -3,
  mark_chimeras = TRUE,
  max_breakpoints = 3L,
  min_parent_ratio = 1.5
)
}
\arguments{
\item{seq_tbl}{(Required). The haplotype sequence table includes sample_id, marker_id, sequence (haplotype sequence), count (haplotype read counts), status (haplotype status), sample, info.}

\item{marker_info}{(Required). The target amplicon file. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence), chrom (chromosome), start (reference sequence start position), end (reference sequence end position).}

\item{output_dir}{(Required). The path to the output file.}

\item{threads}{(Optional). Default 8. The number of processors to use, or NULL to automatically detect and use all available processors.}

\item{min_marker_count}{(Optional). Default 100. The minimum number of reads for each amplicon marker after calling haplotype.}

\item{min_asv_count}{(Optional). Default 5. The minimum number of reads for each haplotype.}

\item{min_asv_freq}{(Optional). Default 0.001. The minimum haplotype frequency.}

\item{min_ident}{(Optional). Default 0.75. The minimum sequence similarity.}

\item{min_ident_z}{(Optional). Default -3. The minimum standardized sequence similarity.}

\item{mark_chimeras}{(Optional). Default TRUE. If true, check chimeric reads.}

\item{max_breakpoints}{(Optional). Default 3. The maximum number of breakpoints chimera reads.}

\item{min_parent_ratio}{min_parent_ratio (Optional). Default 1.5. The minimum ratio that the sequences greater than this value are more abundant than a sequence can be its "parents".}
}
\value{
seq_ann_tbl.rds table: the annotate_seq_tbl table in RDS format includes sample_id, marker_id, sequence (haplotype sequence), count (haplotype read counts), status (haplotype status after annotation), sample, info, ident (sequence similarity), ident_z (standardized sequence similarity).
}
\description{
Haplotype inference for amplicons.
}
