% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indel.R
\name{clean_homopolymers}
\alias{clean_homopolymers}
\title{Clean homopolymers indels}
\usage{
clean_homopolymers(seq_ann_tbl, marker_info, min_homo_rep = 3)
}
\arguments{
\item{seq_ann_tbl}{(Required). The haplotype sequence table includes sample_id, marker_id, sequence (haplotype sequence), count (haplotype read counts).}

\item{marker_info}{(Required). The target amplicon table. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence).}

\item{min_homo_rep}{(Optional). Default 3 (numeric vector). The minimum length of the homopolymer repeats.}
}
\value{
The haplotype sequence with homopolymers indels are changed to be the same as in the reference genome.
}
\description{
Detect indels exist in homopolymers and change the indel to be the same as in the reference genome.
}
\examples{
seq_ann_tbl <- data.frame(
  sample_id = c("S1", "S2", "S3", "S4"),
  marker_id = c("TRAP"),
  sequence = c(
    "TGCTAGTGTTTTTTCAAACAATGCAAGAGAAATTATTAGATTA",
    "TGCTAGTGTTTTCAAACAATGCAAGAGAAATTATTAGATTA",
    "TGCTAGTGTTTTTCAAACAATGCAAGAGAAATTATTAGATTA",
    "TGCTAGTGTTTTTTTTTCAAACAATGCAAGAGAAATTATTAGATTA"
  ),
  count = as.integer(10000)
)

marker_info <- data.frame(
  marker_id = c("TRAP"),
  primer_fwd = c("CTTCTACGTCTTACAAAG"),
  primer_rev = c("CGTGATGATCCTTGCAGA"),
  seq = c("TGCTAGTGTTTTTTCAAACAATGCAAGAGAAATTATTAGATTA")
)

clean_homopolymers(seq_ann_tbl, marker_info)

}
\author{
Jiru Han, Jacob E. Munro, Melanie Bahlo
}
