% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indel.R
\name{clean_terminal_indels}
\alias{clean_terminal_indels}
\title{Clean terminal indels}
\usage{
clean_terminal_indels(seq_ann_tbl, marker_info, terminal_region_len = 1)
}
\arguments{
\item{seq_ann_tbl}{(Required). The haplotype sequence table includes sample_id, marker_id, sequence (haplotype sequence), count (haplotype read counts).}

\item{marker_info}{(Required). The target amplicon table. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence).}

\item{terminal_region_len}{(Optional). Default 1 (numeric vector between 1 and 3). The terminal indel position.}
}
\value{
The haplotype sequence with terminal indels are changed to be the same as in the reference genome.
}
\description{
Change terminal indels to be the same as in the reference genome.
}
\examples{
seq_ann_tbl <- data.frame(
  sample_id = c("S1", "S2", "S3", "S4"),
  marker_id = c("TRAP"),
  sequence = c(
    "TGCTAGTGTTTTTCAAACAATGCAAGAGAAATTATTAGATT",
    "GCTAGTGTTTTTCAAACAATGCAAGAGAAATTATTAGATTA",
    "GCTAGTGTTTTTCAAACAATGCAAGAGAAATTATTAGATT",
    "TGCTAGTGTTTTTTCAAACAATGCAAGAGAAATTATTAGATTA"
  ),
  count = as.integer(10000)
)

marker_info <- data.frame(
  marker_id = c("TRAP"),
  primer_fwd = c("CTTCTACGTCTTACAAAG"),
  primer_rev = c("CGTGATGATCCTTGCAGA"),
  seq = c("TGCTAGTGTTTTTTCAAACAATGCAAGAGAAATTATTAGATTA")
)

clean_terminal_indels(seq_ann_tbl, marker_info)

}
\author{
Jiru Han, Jacob E. Munro, Melanie Bahlo
}
