% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{dada_seq_tbl}
\alias{dada_seq_tbl}
\title{Construct an amplicon sequence variant table (ASV) table}
\usage{
dada_seq_tbl(
  read_table,
  output_dir,
  threads = 1L,
  min_overlap = 10L,
  max_mismatch = 1L,
  trim_overhang = TRUE,
  homo_gap_value = NULL,
  dada_opt = list(HOMOPOLYMER_GAP_PENALTY = -1, MIN_ABUNDANCE = 5),
  marker_info = NULL
)
}
\arguments{
\item{read_table}{The filtered read table includes: sample_id, marker_id, sample, info, reads_1 (the forward read fastq file path), reads_2 (the reverse read fastq file path), n_out (number of reads after filtered), n.}

\item{output_dir}{(Required). The path to the output file.}

\item{threads}{(Optional). The hostnames of workers (as a character vector) or the number of localhost workers (as a positive integer).}

\item{min_overlap}{(Optional). Default 10. The minimum length of the overlap required for merging the forward and reverse reads. When paired-end reads do not overlap, the min_overlap parameter should be set to -1.}

\item{max_mismatch}{(Optional). Default 0. The maximum mismatches allowed in the overlap region.}

\item{trim_overhang}{(Optional). Default FALSE. If TRUE, "overhangs" in the alignment between the forwards and reverse read are trimmed off. "Overhangs" are when the reverse read extends past the start of the forward read, and vice-versa, as can happen when reads are longer than the amplicon and read into the other-direction primer region.}

\item{homo_gap_value}{(Optional). Default NULL (no special homopolymer penalty). The alignment gap penalty within homopolymer regions. Should be negative.}

\item{dada_opt}{(Optional). setDadaOpt sets the default options used by the dada2::dada(...) function for your current session}

\item{marker_info}{(Optional). Default NULL. The target amplicon file. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence), chrom (chromosome), start (reference sequence start position), end (reference sequence end position).}
}
\value{
seq_tbl.rds table: the haplotype sequence table table in RDS format includes sample_id, marker_id, sequence (haplotype sequence), count (haplotype read counts), status (haplotype status), sample, info.
}
\description{
Construct an amplicon sequence variant table (ASV) table.
}
