% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex_marker_only}
\alias{demultiplex_marker_only}
\title{Demultiplex paired-end FASTQ file by marker only}
\usage{
demultiplex_marker_only(sample_read_manifest, marker_info, output_dir, ...)
}
\arguments{
\item{sample_read_manifest}{(Required). The sample reads file. The file should include sample_id, reads_1 (Forward fastq filepaths and fastq filenames have format: SAMPLENAME_R1.fastq.gz), reads_2 (Reverse fastq filepaths and fastq filenames have format: SAMPLENAME_R2.fastq.gz), sample (sample name, can be the same as sample_id), info (e.g., sample type).}

\item{marker_info}{(Required). The target amplicon file. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence), chrom (chromosome), start (reference sequence start position), end (reference sequence end position).}

\item{output_dir}{(Required). The path to the output demultiplexed read table file.}

\item{...}{Other arguments passed to AmpSeqR::demultiplex_reads.}
}
\value{
demultiplex folder: demultiplexed paired-end fastq files.

demultiplex.rds: the demultiplexed table in RDS format which includes sample_id, marker_id, reads_1 (the forward read fastq file path), reads_2 (the reverse read fastq file path), n (number of demultiplexed reads), sample, info.
}
\description{
Demultiplex the paired-end FASTQ reads (demultiplexed by sample) and assigns the sequence to each marker.
}
