% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex_reads}
\alias{demultiplex_reads}
\title{Demultiplex paired-end FASTQ file}
\usage{
demultiplex_reads(
  sample_manifest,
  marker_info,
  reads_1,
  reads_2,
  output_dir,
  output_sub_dir = file.path(output_dir, "demultiplex"),
  use_absolute_paths = TRUE,
  complete_only = TRUE,
  trim_bc = TRUE,
  trim_pr = TRUE,
  trim_right = 0L,
  chunk_n = 1e+06,
  overwrite = TRUE,
  primer_max_mismatch = 3L,
  max_gap_1 = 0L,
  max_gap_2 = 0L,
  degenerate_primers = TRUE,
  suffix_1 = "R1.fastq.gz",
  suffix_2 = "R2.fastq.gz"
)
}
\arguments{
\item{sample_manifest}{(Required). The sample barcodes file. The file should include sample_id, barcode_fwd, barcode_rev, sample (sample name, can be the same as sample_id), info (e.g., sample type).}

\item{marker_info}{(Required). The target amplicon file. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence), chrom (chromosome), start (reference sequence start position), end (reference sequence end position).}

\item{reads_1}{(Required). The file path to the forward fastq file from paired-end sequence data. Compressed file formats such as .fastq.gz are supported.}

\item{reads_2}{(Required). The file path to the reverse fastq file from paired-end sequence data corresponding to those provided to the reads_1 argument. Compressed file formats such as .fastq.gz are supported.}

\item{output_dir}{(Required). The path to the output demultiplexed read table file.}

\item{output_sub_dir}{(Required). The path to the output demultiplexed files.}

\item{use_absolute_paths}{(Required). Default to TRUE. If FALSE, paths to input and output .fastq.gz files will remain relative to the current working directory (useful in case the output is shared with other users so they can possibly continue running the pipeline without running into issues related to different storage mounts).}

\item{complete_only}{(Optional). Default TRUE. If TRUE, only output complete fastq file.}

\item{trim_bc}{(Optional). Default TRUE. If TRUE, trim sample barcodes.}

\item{trim_pr}{(Optional). Default TRUE. If TRUE, trim amplicon primer sequence.}

\item{trim_right}{(Optional). Default 0. The number of nucleotides to remove from the end of each read.}

\item{chunk_n}{(Optional). Default 1e6. For ShortRead::FastqSampler, the size of the sample (number of records) to be drawn. For ShortRead::FastqStreamer a numeric(1) (set to 1e6 when n is missing) providing the number of successive records to be returned on each yield, or an IRanges-class delimiting the (1-based) indicies of records returned by each yield; entries in n must have non-zero width and must not overlap.}

\item{overwrite}{(Optional). Default TRUE. If TRUE, overwrite output files already exist.}

\item{primer_max_mismatch}{(Optional). Default 3. The maximum number of mismatches of the amplicon primer sequence.}

\item{max_gap_1}{(Optional). Default 0. The maximum number of bases between left end of read and start of barcode sequence.}

\item{max_gap_2}{(Optional). Default 0. The maximum number of bases between the barcode and the marker primer sequence.}

\item{degenerate_primers}{(Optional). Default TRUE. If TRUE (the default), an IUPAC ambiguity code in the pattern can only match the same code in the subject, and vice versa. If FALSE, an IUPAC ambiguity code in the pattern can match any letter in the subject that is associated with the code, and vice versa. See Biostrings::vmatchPattern for more information.}

\item{suffix_1}{(Optional). Default "R1.fastq.gz". Output forward fastq file name extension.}

\item{suffix_2}{(Optional). Default "R2.fastq.gz". Output reverse fastq file name extension.}
}
\value{
demultiplex folder: demultiplexed paired-end fastq files.

demultiplex.rds: the demultiplexed table in RDS format which includes sample_id, marker_id, reads_1 (the forward read fastq file path), reads_2 (the reverse read fastq file path), n (number of demultiplexed reads), sample, info.
}
\description{
Demultiplex the raw paired-end FASTQ reads with mixed samples and trim the sample barcodes and target amplicon primer sequences and assigns the sequence to each sample each marker.
}
\examples{

example_data <- get_ampseqr_example_data()

# Set the directory
run_dir <- "runs"
dir.create(run_dir)

demultiplexed <- demultiplex_reads(
  sample_manifest = example_data$sample_manifest,
  marker_info = example_data$marker_info,
  reads_1 = example_data$reads_1,
  reads_2 = example_data$reads_2,
  output_dir = run_dir,
  output_sub_dir = file.path(run_dir, "demultiplex")
)

}
