% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{downsample_reads}
\alias{downsample_reads}
\title{Downsample read counts}
\usage{
downsample_reads(
  read_table,
  output_dir,
  output_sub_dir,
  use_absolute_paths = TRUE,
  min_read_count = 1000,
  n_sample = 10000,
  seed = 1L,
  threads = 2L,
  count_col = NULL
)
}
\arguments{
\item{read_table}{(Required). The filtered read table includes: sample_id, marker_id, n (number of demultiplexed reads), sample, info, reads_1 (the filtered and trimmed forward fastq file path), reads_2 (the filtered and trimmed reverse fastq file path), n_in (number of reads before filtering and trimming), n_out (number of reads after filtering and trimming).}

\item{output_dir}{(Required). The path to the output downsampled read table file.}

\item{output_sub_dir}{(Required). The path to the output downsampled files.}

\item{use_absolute_paths}{(Required). Default to TRUE. If FALSE, paths to input and output .fastq.gz files will remain relative to the current working directory (useful in case the output is shared with other users so they can possibly continue running the pipeline without running into issues related to different storage mounts).}

\item{min_read_count}{(Optional). Default 1000. The minimum number of reads per sample per marker.}

\item{n_sample}{(Optional). Default 10000. Downsamples an exact number of reads from paired end fastq files.}

\item{seed}{(Optional). Default 1. Random seed for reproducible downsampling.}

\item{threads}{(Optional). Default 2. The hostnames of workers (as a character vector) or the number of localhost workers (as a positive integer).}

\item{count_col}{(Optional). Default NULL. The specified column to downsample reads.}
}
\value{
A downsampled table in RDS format includes sample_id, marker_id, sample, info, reads_1 (the downsampled forward fastq file path), reads_2 (the downsampled reverse fastq file path), n (number of reads after downsampling).
}
\description{
This function downsamples an exact number of reads from paired end fastq files.
}
