% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate the HTML report including various summary data visualizations}
\usage{
generate_report(
  sample_manifest,
  marker_info,
  demultiplexed,
  flt_reads,
  sub_reads,
  seq_ann_tbl,
  seq_flt_tbl,
  report_output_dir,
  browse_report = TRUE
)
}
\arguments{
\item{sample_manifest}{(Required). The sample barcodes file. The file should include sample_id, barcode_fwd, barcode_rev, sample (sample name, can be the same as sample_id), info (e.g., sample type).}

\item{marker_info}{(Required). The target amplicon file. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence), chrom (chromosome), start (reference sequence start position), end (reference sequence end position).}

\item{demultiplexed}{(Required). The demultiplexed read table includes: sample_id, marker_id, reads_1 (the forward read fastq file path), reads_2 (the reverse read fastq file path), n (number of demultiplexed reads), sample, info.}

\item{flt_reads}{(Required). The filtered read table includes: sample_id, marker_id, n (number of demultiplexed reads), sample, info, reads_1 (the filtered and trimmed forward fastq file path), reads_2 (the filtered and trimmed reverse fastq file path), n_in (number of reads before filtering and trimming), n_out (number of reads after filtering and trimming).}

\item{sub_reads}{(Required). The downsampled read table includes: sample_id, marker_id, sample, info, reads_1 (the downsampled forward fastq file path), reads_2 (the downsampled reverse fastq file path), n (number of reads after downsampling).}

\item{seq_ann_tbl}{(Required). The amplicon haplotype sequence table includes: sample_id, marker_id, sequence (haplotype sequence), count (read counts), status (haplotype status), sample, info, ident (sequence similarity), ident_z (standardized sequence similarity).}

\item{seq_flt_tbl}{(Required). The final amplicon haplotype sequence table includes: sample_id, marker_id, sequence (haplotype sequence), count (read counts), masked (TRUE indicates the sequence contains variants that might be errors, and replaces the nucleotide with the reference genome nucleotide), status (haplotype status, all pass), ident (sequence similarity), haplotype (a named panel of haplotypes for each marker), frequency (within-sample haplotype frequency for each marker), sample, info.}

\item{report_output_dir}{(Required). The path to the output report file.}

\item{browse_report}{(Optional). Default TRUE. Logical indicating whether the htmlFile should be opened in a web browser.}
}
\value{
The HTML report including various summary data visualizations.
}
\description{
Generate the HTML report including various summary data visualizations.
}
