% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_ampseqr_example_data}
\alias{get_ampseqr_example_data}
\title{Get AmpSeqR extdata}
\usage{
get_ampseqr_example_data()
}
\value{
List of the example data.
}
\description{
The example data that is required for AmpSeqR. The input files are the standard paired-end FASTQ files provided by the common NGS sequencing platforms, as well as sample barcodes and target amplicon details. The sample barcodes file should include sample_id, barcode_fwd, barcode_rev, sample (sample name, can be the same as sample_id), info (e.g., sample type).  The target amplicon details file should include marker_id, primer_fwd, primer_rev, seq (reference sequence), chrom (chromosome), start (reference sequence start position), end (reference sequence end position).
}
\examples{
example_data <- get_ampseqr_example_data()
}
