% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quality.R
\name{plot_quality}
\alias{plot_quality}
\title{Plot quality profile of fastq files}
\usage{
plot_quality(data)
}
\arguments{
\item{data}{(Required). The demultiplexed read table includes: sample_id, marker_id, reads_1 (the forward read fastq file path), reads_2 (the reverse read fastq file path), n (number of demultiplexed reads), sample, info.}
}
\value{
Generate a visual summary plot of the distribution of quality scores of forward and reverse sequence position for all fastq files.
}
\description{
Visualize the quality profiles of the forward and reverse sequence position.
}
