% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_filter.R
\name{sequence_filter}
\alias{sequence_filter}
\title{Optimize amplicon haplotype calling}
\usage{
sequence_filter(
  seq_ann_tbl,
  sample_manifest,
  marker_info,
  output_dir,
  vcf_output_dir,
  min_homo_rep = 3,
  terminal_region_len = 1,
  max_sm_miss = 0.5,
  max_marker_miss = 0.5,
  sample_med_He = 0,
  n_alleles = 3,
  var_maf = 0.001,
  var_he = 0.001
)
}
\arguments{
\item{seq_ann_tbl}{(Required). The amplicon haplotype sequence table includes: sample_id, marker_id, sequence (haplotype sequence), count (read counts), status (haplotype status), sample, info, ident (sequence similarity), ident_z (standardized sequence similarity).}

\item{sample_manifest}{(Required). The sample barcodes file. The file should include sample_id, barcode_fwd, barcode_rev, sample (sample name, can be the same as sample_id), info (e.g., sample type).}

\item{marker_info}{(Required). The target amplicon file. The file should include marker_id, primer_fwd, primer_rev, seq (reference sequence), chrom (chromosome), start (reference sequence start position), end (reference sequence end position).}

\item{output_dir}{(Required). The path to the output sequence table file.}

\item{vcf_output_dir}{(Required). The path to the output vcf files.}

\item{min_homo_rep}{(Optional). Default 3. The minimum size for homopolymer in reference sequence.}

\item{terminal_region_len}{(Optional). Default 1. The length of the terminal region. The maximum value is 3.}

\item{max_sm_miss}{(Optional). Default 0.5. The maximum fraction of missing data for samples.}

\item{max_marker_miss}{(Optional). Default 0.5. The maximum fraction of missing data for amplicon markers.}

\item{sample_med_He}{(Optional). Default 0. The maximum median expected heterozygosity (a measure of genetic variation within populations) for the variance of all samples in the dataset.}

\item{n_alleles}{(Optional). Default 3. The maximum number of alleles at a locus.}

\item{var_maf}{(Optional). Default 0.001. The minimum minor allele frequency (MAF) of the variant.}

\item{var_he}{(Optional). Default 0.001. The minimum expected heterozygosity of the variant.}
}
\value{
An amplicon haplotype sequence table in RDS format which includes sample_id, marker_id, sequence (haplotype sequence), count (read counts), masked (TRUE indicates the sequence contains variants that might be errors, and replaces the nucleotide with the reference genome nucleotide), status (haplotype status, all pass), ident (sequence similarity), haplotype (a named panel of haplotypes for each marker), frequency (within-sample haplotype frequency for each marker), sample, info.
}
\description{
This function optimize amplicon haplotype calling and generate an amplicon haplotype sequence table.
}
